requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
celery==5.3.6
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
pymssql==2.2.11  # For connecting to HitTrax MS SQL database
psycopg2-binary==2.9.9  # For connecting to HitTrax PostgreSQL database
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
celery==5.3.6
pymssql==2.2.11
pandas==2.1.4
python-dotenv==1.0.0
sqlalchemy
plotly
scipy